+++ tests/bootstrap.php
@@ -8,3 +8,4 @@
      */
     require_once __DIR__ . '/../vendor/autoload.php';
     require_once __DIR__ . '/../JkPublishPagesRules.php';
+    require_once __DIR__ . '/Integration/IntegrationTestCase.php';